cambios 2da entrega

+++ Proyecto/Proyecto/View/Brands.cs
@@ -43,7 +43,7 @@ namespace Proyecto.View
                 //verify the camps is not null
                 if (String.IsNullOrEmpty(txtCode.Text) || String.IsNullOrEmpty(txtDescription.Text))
                 {
-                    MessageBox.Show("Todos los campos son requeridos");
+                    MessageBox.Show("Todos los campos son requeridos", "Error", MessageBoxButtons.OK, MessageBoxIcon.Warning);
 
                 }
                 else
@@ -64,7 +64,7 @@ namespace Proyecto.View
             }
             catch
             {
-                MessageBox.Show("Error verifique los datos insertados");
+                MessageBox.Show("Problema al crear Marca", "Error", MessageBoxButtons.OK, MessageBoxIcon.Warning);
             }
             return oBrandE;
         }
@@ -76,27 +76,40 @@ namespace Proyecto.View
             {
                 if (oBreandD.insertBrand(oBrandE))
                 {
-                    MessageBox.Show("Marca insertada correctamente");
+                    MessageBox.Show("Marca insertado correctamente","Inserción",MessageBoxButtons.OK,MessageBoxIcon.Information);
                     chargeDataGrid();
+                    cleanForm();
+                }
+                else
+                {
+                    MessageBox.Show("Error al insertar marca: " + oBreandD.ErrorMsg,"Inserción",MessageBoxButtons.OK,MessageBoxIcon.Error);
                 }
             }
             else
             {
-                MessageBox.Show("Error al insertar marca: " + oBreandD.ErrorMsg);
+                MessageBox.Show("Error al crear marca", "Inserción", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
         }
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            string code = dataGridView1.CurrentRow.Cells["Code"].Value.ToString();
-            if (oBreandD.deleteBrand(code))
+            if (MessageBox.Show("Desea eliminar la marca", "Eliminación", MessageBoxButtons.YesNo,MessageBoxIcon.Question) == DialogResult.Yes)
             {
-                MessageBox.Show("Marca eliminada exitosamente");
-                chargeDataGrid();
+                string code = dataGridView1.CurrentRow.Cells["Code"].Value.ToString();
+                if (oBreandD.deleteBrand(code))
+                {
+                    MessageBox.Show("Marca eliminado correctamente", "Eliminación", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    chargeDataGrid();
+                    cleanForm();
+                }
+                else
+                {
+                    MessageBox.Show("Error al eliminar marca: " + oBreandD.ErrorMsg, "Eliminación", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
             }
             else
             {
-                MessageBox.Show("Error al elimimar marca: " + oBreandD.ErrorMsg);
+                MessageBox.Show("Eliminacion cancelada", "Eliminación", MessageBoxButtons.OK, MessageBoxIcon.Information);
             }
         }
 
@@ -114,14 +127,29 @@ namespace Proyecto.View
             {
                 if (oBreandD.updateBrand(oBrandE))
                 {
-                    MessageBox.Show("Marca modificada correctamente");
+                    MessageBox.Show("Marca modificada correctamente", "Modificación", MessageBoxButtons.OK, MessageBoxIcon.Information);
                     chargeDataGrid();
+                    cleanForm();
+                }
+                else
+                {
+                    MessageBox.Show("Error al modificar marca: " + oBreandD.ErrorMsg, "Modificación", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
             }
             else
             {
-                MessageBox.Show("Error al modificar marca: " + oBreandD.ErrorMsg);
+                MessageBox.Show("Error al crear marca", "Modificación", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
+        }
+
+        private void cleanForm()
+        {
+            txtCode.Text = "";
+            txtDescription.Text = "";
+        }
+        private void btnClean_Click(object sender, EventArgs e)
+        {
+            cleanForm();
         }
     }
 }
